+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 ## _v.0.1.0_
 ## Implementation plan
 
-* [ ] Separate functionality to Encode and Decode
+* [x] Separate functionality to Encode and Decode
 * [ ] Pop-up menu for left-clicking on text widget to save to file
 * [ ] Add ability to select modes to Caesar: alphabet and ASCII
 * [ ] Add information in window Help